Configurable beautified vinit with IRQs.
Emulators are treated and configurable like other childs.

Multiple emulation contexts can be assigned to one type of
emulator via configuration.

Any emulation context is assigned to one emulator via
configuration.

IRQs can be emulated by an emulator.

+++ base/include/base/pager.h
@@ -77,8 +77,8 @@ namespace Genode {
 			 * \param  read   If the instruction has implied a memory
 			 *                read access, this contains the resulting value
 			 */
-			void instruction_processed(unsigned long * const instr,
-			                           unsigned long const read);
+			void instruction_processed(unsigned * const instr,
+			                           unsigned const read);
 
 			void wake_up()
 			{
